requirements.txt
pandas==2.2.0
mlflow==2.2.2
notebook
numpy==1.26.4
matplotlib==3.8.2
python-dotenv==1.0.1
scikit-learn
seaborn==0.13.2
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
dvc[s3]
python-box
dvclive>=1.0
fastapi[all]
uvicorn